feat(opensearch_helper): create remove legacy id script
